add shell support everywhere
