+++ sdk/src/beta9/abstractions/base/runner.py
@@ -27,6 +27,7 @@ from ...clients.gateway import (
     SecretVar,
 )
 from ...clients.gateway import TaskPolicy as TaskPolicyProto
+from ...clients.shell import ShellServiceStub
 from ...config import ConfigContext, SDKSettings, get_config_context, get_settings
 from ...env import called_on_import
 from ...sync import FileSyncer, SyncEventHandler
@@ -144,6 +145,7 @@ class RunnerAbstraction(BaseAbstraction):
             self._map_callable_to_attr(attr="on_start", func=on_start)
 
         self._gateway_stub: Optional[GatewayServiceStub] = None
+        self._shell_stub: Optional[ShellServiceStub] = None
         self.syncer: FileSyncer = FileSyncer(self.gateway_stub)
         self.settings: SDKSettings = get_settings()
         self.config_context: ConfigContext = get_config_context()
@@ -211,6 +213,16 @@ class RunnerAbstraction(BaseAbstraction):
     @gateway_stub.setter
     def gateway_stub(self, value) -> None:
         self._gateway_stub = value
+
+    @property
+    def shell_stub(self) -> ShellServiceStub:
+        if not self._shell_stub:
+            self._shell_stub = ShellServiceStub(self.channel)
+        return self._shell_stub
+
+    @shell_stub.setter
+    def shell_stub(self, value) -> None:
+        self._shell_stub = value
 
     def _parse_cpu_to_millicores(self, cpu: Union[float, str]) -> int:
         """
